+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.16
+
+### Patch Changes
+
+- Update CrudItem typing
+
 ## 0.0.15
 
 ### Patch Changes